refactors goodreads service to google service;
