--- components/goodreads/index.js
@@ -1,5 +1,0 @@
-const goodreadRoutes = require('./goodreadRoutes');
-
-module.exports = (server) => {
-  server.use('/api/goodreads', goodreadRoutes)
-}
