+++ components/google/googleRoutes.js
@@ -1,5 +1,5 @@
 const router = require('express').Router();
-const controllers = require('./goodreadControllers');
+const controllers = require('./googleControllers');
 
 // SEARCH BOOKS BY TITLE, AUTHOR, OR ISBN
 router.get('/search', controllers.searchBooks);
