+++ components/google/index.js
@@ -1,0 +1,5 @@
+const googleRoutes = require('./googleRoutes');
+
+module.exports = server => {
+  server.use('/api/google', googleRoutes);
+};
