+++ components/index.js
@@ -2,18 +2,17 @@ const router = require('express').Router();
 const { lstatSync, readdirSync } = require('fs');
 const { join } = require('path');
 
-const isDirectory = (__dirname) => lstatSync(__dirname).isDirectory() // checks if files are directories  true || false
+const isDirectory = __dirname => lstatSync(__dirname).isDirectory(); // checks if files are directories  true || false
 const getDirectories = () =>
   readdirSync(__dirname) // files inside target directory
     .map(
-      (name) => join(__dirname, name) //index.js & user
+      name => join(__dirname, name) //index.js & user
     )
-    .filter(isDirectory)
+    .filter(isDirectory);
 
-const directoryArr = getDirectories() // Array full of all directory paths
+const directoryArr = getDirectories(); // Array full of all directory paths
 
-
-module.exports = (server) => {
+module.exports = server => {
   require('./posts')(server);
   require('./auth')(server);
   require('./services')(server);
@@ -21,7 +20,6 @@ module.exports = (server) => {
   require('./users')(server);
   require('./comments')(server);
   require('./images')(server);
-  require('./notifications')(server)
-  require('./goodreads')(server)
-}
-
+  require('./notifications')(server);
+  require('./google')(server);
+};
